fix: remove duplicated breadcrumbs of asset inventory page (#2807)
* fix: remove duplicated breadcrumbs of asset inventory page


* refactor: separate project container and lnb


* fix: remove duplicated breadcrumbs from all pages


---------
