+++ apps/web/src/common/modules/page-layouts/VerticalPageLayout.vue
@@ -14,7 +14,9 @@
                  :style="{ height: globalUIGetters.appBodyHeight }"
             >
                 <portal-target name="page-top-notification" />
-                <div class="header">
+                <div class="header"
+                     :class="{'without-breadcrumbs': !breadcrumbs.length}"
+                >
                     <p-breadcrumbs v-if="breadcrumbs.length"
                                    :routes="breadcrumbs"
                                    :copiable="copiable"
@@ -101,6 +103,9 @@ export default {
     .header {
         @apply flex justify-between;
         padding: 1.5rem 1.5rem 0.25rem 1.5rem;
+        &.without-breadcrumbs {
+            padding: 0;
+        }
     }
 
     .page-contents {
