+++ apps/web/src/services/administration/routes/admin/routes.ts
@@ -3,6 +3,7 @@ import type { RouteConfig } from 'vue-router';
 import { makeAdminRouteName } from '@/router/helpers/route-helper';
 
 import { MENU_ID } from '@/lib/menu/config';
+import { MENU_INFO_MAP } from '@/lib/menu/menu-info';
 
 import { ADMINISTRATION_ROUTE } from '@/services/administration/routes/route-constant';
 
@@ -21,7 +22,10 @@ const WorkspacesPage = () => import('@/services/administration/pages/admin/Admin
 const adminAdministrationRoutes: RouteConfig = {
     path: 'administration',
     name: makeAdminRouteName(ADMINISTRATION_ROUTE._NAME),
-    meta: { menuId: MENU_ID.ADMINISTRATION },
+    meta: {
+        menuId: MENU_ID.ADMINISTRATION,
+        translationId: MENU_INFO_MAP[MENU_ID.ADMINISTRATION].translationId,
+    },
     redirect: () => ({
         name: makeAdminRouteName(ADMINISTRATION_ROUTE.IAM._NAME),
     }),
@@ -30,7 +34,10 @@ const adminAdministrationRoutes: RouteConfig = {
         {
             path: 'iam',
             name: makeAdminRouteName(ADMINISTRATION_ROUTE.IAM._NAME),
-            meta: { menuId: MENU_ID.IAM },
+            meta: {
+                menuId: MENU_ID.IAM,
+                translationId: MENU_INFO_MAP[MENU_ID.IAM].translationId,
+            },
             redirect: () => ({
                 name: makeAdminRouteName(ADMINISTRATION_ROUTE.IAM.USER._NAME),
             }),
@@ -39,18 +46,29 @@ const adminAdministrationRoutes: RouteConfig = {
                 {
                     path: 'user',
                     name: makeAdminRouteName(ADMINISTRATION_ROUTE.IAM.USER._NAME),
-                    meta: { lnbVisible: true, menuId: MENU_ID.USER },
+                    meta: {
+                        lnbVisible: true,
+                        menuId: MENU_ID.USER,
+                        translationId: MENU_INFO_MAP[MENU_ID.USER].translationId,
+                    },
                     component: UserMainPage as any,
                 },
                 {
                     path: 'app',
                     name: makeAdminRouteName(ADMINISTRATION_ROUTE.IAM.APP._NAME),
-                    meta: { lnbVisible: true, menuId: MENU_ID.APP },
+                    meta: {
+                        lnbVisible: true,
+                        menuId: MENU_ID.APP,
+                        translationId: MENU_INFO_MAP[MENU_ID.APP].translationId,
+                    },
                     component: AppMainPage as any,
                 },
                 {
                     path: 'role',
-                    meta: { menuId: MENU_ID.ROLE },
+                    meta: {
+                        menuId: MENU_ID.ROLE,
+                        translationId: MENU_INFO_MAP[MENU_ID.ROLE].translationId,
+                    },
                     component: { template: '<router-view />' },
                     children: [
                         {
@@ -80,7 +98,10 @@ const adminAdministrationRoutes: RouteConfig = {
         {
             path: 'preference',
             name: makeAdminRouteName(ADMINISTRATION_ROUTE.PREFERENCE._NAME),
-            meta: { menuId: MENU_ID.PREFERENCE },
+            meta: {
+                menuId: MENU_ID.PREFERENCE,
+                translationId: MENU_INFO_MAP[MENU_ID.PREFERENCE].translationId,
+            },
             redirect: () => ({
                 name: makeAdminRouteName(ADMINISTRATION_ROUTE.PREFERENCE.DOMAIN_SETTINGS._NAME),
             }),
@@ -89,13 +110,21 @@ const adminAdministrationRoutes: RouteConfig = {
                 {
                     path: 'domain-settings',
                     name: makeAdminRouteName(ADMINISTRATION_ROUTE.PREFERENCE.DOMAIN_SETTINGS._NAME),
-                    meta: { lnbVisible: true, menuId: MENU_ID.DOMAIN_SETTINGS },
+                    meta: {
+                        lnbVisible: true,
+                        menuId: MENU_ID.DOMAIN_SETTINGS,
+                        translationId: MENU_INFO_MAP[MENU_ID.DOMAIN_SETTINGS].translationId,
+                    },
                     component: DomainSettingsPage,
                 },
                 {
                     path: 'workspaces',
                     name: makeAdminRouteName(ADMINISTRATION_ROUTE.PREFERENCE.WORKSPACES._NAME),
-                    meta: { lnbVisible: true, menuId: MENU_ID.WORKSPACES },
+                    meta: {
+                        lnbVisible: true,
+                        menuId: MENU_ID.WORKSPACES,
+                        translationId: MENU_INFO_MAP[MENU_ID.WORKSPACES].translationId,
+                    },
                     component: WorkspacesPage,
                 },
             ],
