+++ apps/web/src/services/asset-inventory/routes/admin/routes.ts
@@ -5,6 +5,7 @@ import { upperCase } from 'lodash';
 import { makeAdminRouteName } from '@/router/helpers/route-helper';
 
 import { MENU_ID } from '@/lib/menu/config';
+import { MENU_INFO_MAP } from '@/lib/menu/menu-info';
 
 import { ASSET_INVENTORY_ROUTE } from '@/services/asset-inventory/routes/route-constant';
 
@@ -29,12 +30,15 @@ const CollectorDetailPage = () => import('@/services/asset-inventory/pages/Colle
 const adminAssetInventoryRoute: RouteConfig = {
     path: 'asset-inventory',
     name: makeAdminRouteName(ASSET_INVENTORY_ROUTE._NAME),
-    meta: { menuId: MENU_ID.ASSET_INVENTORY },
+    meta: { menuId: MENU_ID.ASSET_INVENTORY, translationId: MENU_INFO_MAP[MENU_ID.ASSET_INVENTORY].translationId },
     component: AssetInventoryContainer,
     children: [
         {
             path: 'cloud-service',
-            meta: { menuId: MENU_ID.CLOUD_SERVICE },
+            meta: {
+                menuId: MENU_ID.CLOUD_SERVICE,
+                translationId: MENU_INFO_MAP[MENU_ID.CLOUD_SERVICE].translationId,
+            },
             component: { template: '<router-view />' },
             children: [
                 {
@@ -80,12 +84,12 @@ const adminAssetInventoryRoute: RouteConfig = {
         {
             path: 'server',
             name: makeAdminRouteName(ASSET_INVENTORY_ROUTE.SERVER._NAME),
-            meta: { lnbVisible: true, menuId: MENU_ID.SERVER },
+            meta: { lnbVisible: true, menuId: MENU_ID.SERVER, translationId: MENU_INFO_MAP[MENU_ID.SERVER].translationId },
             component: ServerPage as any,
         },
         {
             path: 'collector',
-            meta: { menuId: MENU_ID.COLLECTOR },
+            meta: { menuId: MENU_ID.COLLECTOR, translationId: MENU_INFO_MAP[MENU_ID.COLLECTOR].translationId },
             component: { template: '<router-view />' },
             children: [
                 {
